+++ common/common-test-setup/src/base_setup/contract_endpoints.rs
@@ -1,0 +1,73 @@
+use multiversx_sc_scenario::{
+    api::StaticApi,
+    imports::{ManagedBuffer, MultiValueEncoded, TestAddress},
+    ReturnsHandledOrError, ScenarioTxRun,
+};
+use proxies::{fee_market_proxy::FeeMarketProxy, header_verifier_proxy::HeaderverifierProxy};
+use structs::fee::FeeStruct;
+
+use crate::{
+    base_setup::init::BaseSetup,
+    constants::{FEE_MARKET_ADDRESS, HEADER_VERIFIER_ADDRESS, OWNER_ADDRESS},
+};
+
+impl BaseSetup {
+    pub fn register_operation(
+        &mut self,
+        caller: TestAddress,
+        signature: ManagedBuffer<StaticApi>,
+        hash_of_hashes: &ManagedBuffer<StaticApi>,
+        operations_hashes: MultiValueEncoded<StaticApi, ManagedBuffer<StaticApi>>,
+    ) {
+        self.world
+            .tx()
+            .from(caller)
+            .to(HEADER_VERIFIER_ADDRESS)
+            .typed(HeaderverifierProxy)
+            .register_bridge_operations(
+                signature,
+                hash_of_hashes,
+                ManagedBuffer::new(),
+                ManagedBuffer::new(),
+                operations_hashes,
+            )
+            .run();
+    }
+
+    pub fn set_fee_during_setup_phase(
+        &mut self,
+        fee_struct: FeeStruct<StaticApi>,
+        error_message: Option<&str>,
+    ) {
+        let response = self
+            .world
+            .tx()
+            .from(OWNER_ADDRESS)
+            .to(FEE_MARKET_ADDRESS)
+            .typed(FeeMarketProxy)
+            .set_fee_during_setup_phase(fee_struct)
+            .returns(ReturnsHandledOrError::new())
+            .run();
+
+        self.assert_expected_error_message(response, error_message);
+    }
+
+    pub fn set_fee(
+        &mut self,
+        hash_of_hashes: &ManagedBuffer<StaticApi>,
+        fee_struct: Option<FeeStruct<StaticApi>>,
+        error_message: Option<&str>,
+    ) {
+        let response = self
+            .world
+            .tx()
+            .from(OWNER_ADDRESS)
+            .to(FEE_MARKET_ADDRESS)
+            .typed(FeeMarketProxy)
+            .set_fee(hash_of_hashes, fee_struct.unwrap())
+            .returns(ReturnsHandledOrError::new())
+            .run();
+
+        self.assert_expected_error_message(response, error_message);
+    }
+}
