+++ common/common-test-setup/src/base_setup/mod.rs
@@ -1,0 +1,6 @@
+pub mod checks;
+pub mod complete_setup_phase;
+pub mod contract_endpoints;
+pub mod deploy;
+pub mod helpers;
+pub mod init;
