+++ mvx-esdt-safe/tests/mvx_esdt_safe_blackbox_setup.rs
@@ -1,9 +1,9 @@
+use common_test_setup::base_setup::init::{AccountSetup, BaseSetup, RegisterTokenArgs};
 use common_test_setup::constants::{
     ESDT_SAFE_ADDRESS, FEE_TOKEN, FIRST_TEST_TOKEN, HEADER_VERIFIER_ADDRESS,
     MVX_ESDT_SAFE_CODE_PATH, ONE_HUNDRED_MILLION, OWNER_ADDRESS, OWNER_BALANCE, SECOND_TEST_TOKEN,
     USER_ADDRESS,
 };
-use common_test_setup::{AccountSetup, BaseSetup, RegisterTokenArgs};
 use multiversx_sc::{
     imports::OptionalValue,
     types::{
