+++ sov-esdt-safe/tests/sov_esdt_safe_blackbox_setup.rs
@@ -7,11 +7,11 @@ use multiversx_sc_scenario::{
     api::StaticApi, ReturnsHandledOrError, ReturnsLogs, ScenarioTxRun, ScenarioTxWhitebox,
 };
 
+use common_test_setup::base_setup::init::{AccountSetup, BaseSetup};
 use common_test_setup::constants::{
     ESDT_SAFE_ADDRESS, FEE_MARKET_ADDRESS, FEE_TOKEN, FIRST_TEST_TOKEN, ONE_HUNDRED_MILLION,
     OWNER_ADDRESS, OWNER_BALANCE, SECOND_TEST_TOKEN, SOV_ESDT_SAFE_CODE_PATH, USER_ADDRESS,
 };
-use common_test_setup::{AccountSetup, BaseSetup};
 use proxies::sov_esdt_safe_proxy::SovEsdtSafeProxy;
 use sov_esdt_safe::SovEsdtSafe;
 use structs::{
