+++ sovereign-forge/tests/sovereign_forge_blackbox_setup.rs
@@ -1,9 +1,9 @@
 use common_test_setup::{
+    base_setup::init::{AccountSetup, BaseSetup},
     constants::{
         CHAIN_FACTORY_SC_ADDRESS, ESDT_SAFE_ADDRESS, OWNER_ADDRESS, OWNER_BALANCE,
         SOVEREIGN_FORGE_SC_ADDRESS, TOKEN_HANDLER_SC_ADDRESS,
     },
-    AccountSetup, BaseSetup,
 };
 use multiversx_sc::{
     imports::OptionalValue,
